requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
ragflow-sdk
openai>=1.55.0